Change product quantity to float in getProductos
Updated the 'cantidad' field in the getProductos function to cast quantity as a float instead of an int. This allows for handling fractional product quantities.

+++ configurar/creditos_por_cliente.php
@@ -76,7 +76,7 @@ function getProductos($orderId)
 
         $productos[] = [
             'id'       => (int)$row['product_id'],
-            'cantidad' => (int)$row['quantity'],
+            'cantidad' => (float)$row['quantity'],
             'datos'    => $datos,
         ];
     }
